build_requirements.txt
# Enhanced Build Requirements for Afarensis Enterprise v2.1
# Includes all dependencies for Advanced Search & Collaborative Review Features

# Core build tools
pyinstaller>=6.2.0
pydantic-settings>=2.1.0
setuptools>=69.0.0
wheel>=0.42.0

# Enhanced AI and Search Dependencies (NEW in v2.1)
sentence-transformers>=2.2.2
transformers>=4.35.2
torch>=2.1.1
numpy>=1.24.4
scipy>=1.11.4
faiss-cpu>=1.7.4
anthropic>=0.8.1
openai>=1.3.7

# Real-time Collaboration Dependencies (NEW in v2.1)  
websockets>=12.0
python-socketio>=5.10.0

# Citation Analysis & Network Processing (NEW in v2.1)
networkx>=3.2.1
matplotlib>=3.8.2

# Natural Language Processing (NEW in v2.1)
spacy>=3.7.2
nltk>=3.8.1
textblob>=0.17.1

# Core application dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
alembic>=1.13.0
asyncpg>=0.29.0

# HTTP and API integration
aiohttp>=3.9.1
httpx>=0.25.2
requests>=2.31.0
beautifulsoup4>=4.12.2

# Background tasks and caching
redis>=5.0.1
celery>=5.3.4

# Security and authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
cryptography>=41.0.8

# Configuration and utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
click>=8.1.7

# Document processing
PyPDF2>=3.0.1
python-docx>=1.1.0

# Data processing
pandas>=2.1.4
python-dateutil>=2.8.2

# GUI and visualization
pillow>=10.1.0
tkinter  # Included with Python

# Platform-specific dependencies
pywin32>=306 ; sys_platform == "win32"
pywin32-ctypes>=0.2.2 ; sys_platform == "win32"
